Add game logic sensitivity for analog yaw & pitch
Add options to change how fast the analog sensitivity is for yaw & pitch.

Why not use the axis sensitivity options? Because they can change how the deadzones and input response preform, and you have to balance several options. In my opinion, those sensitivity options should be for full-on stick calibration (if your stick is not outputting the full range, for instance). But how fast the player should be able to look around is up to the user's taste, regardless of their stick settings.

Default analog pitch sensitivity has been drastically lowered, from 2048 (x1.6 yaw sensitivity) to 832 (x0.65 of yaw sensitivity). I don't think I've ever heard of any FPS game on controller before that makes pitch more sensitive than yaw... it is either the same or lower, because players typically want to stay looking in the horizon, and analog stick isn't precise enough to treat them evenly (let alone more sensitive). IMO this is a lot more playable, but I can change the default back to match older versions if that is preferred.